Reorganized Header
Reorganized the header and altered the scss. Instead of all being navigation, the header now consists of three parts: a titlebar, a navigation list, and a wrapper class called simply header. The navigation list is positioned within the header and is placed above the titlebar when the screen is larger than a tablet, but below the titlebar when it is smaller.